requirements.txt
Django~=3.0.5
django-cors-headers==3.1.1
django-environ==0.4.5
djangorestframework~=3.11.0
inflection==0.3.1
psycopg2==2.7.1
six==1.15.0
environ==1.0
pyyaml==6.0
uritemplate==4.1.1